Add utils.net
