+++ utils/__init__.py
@@ -1,2 +1,3 @@
+from .net import *  # noqa: F401, F403
 from .types import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
