+++ vsrepo/__init__.py
@@ -24,7 +24,6 @@ import argparse
 import base64
 import binascii
 import csv
-import email.utils
 import glob
 import hashlib
 import importlib.util as imputil
@@ -35,12 +34,15 @@ import re
 import subprocess
 import sys
 import tempfile
-import urllib.request
 import zipfile
+from email.utils import formatdate, mktime_tz, parsedate_tz
 from pathlib import Path
-from typing import Dict, Iterator, List, MutableMapping, Optional, Tuple, cast
+from typing import Dict, Iterator, List, MutableMapping, Optional, Tuple
+from urllib.error import HTTPError
+from urllib.request import Request, urlopen
 
 from utils import BoundVSPackageRelT, VSPackage, VSPackageRel, VSPackages
+from utils.net import fetch_url_cached
 from utils.types import VSPackageType
 
 try:
@@ -49,10 +51,6 @@ except ImportError:
     print('{} is only supported on Windows.'.format(__file__))
     sys.exit(1)
 
-try:
-    import tqdm  # type: ignore
-except ImportError:
-    pass
 
 bundled_api3_plugins = {
     'com.vapoursynth.avisource', 'com.vapoursynth.eedi3', 'com.vapoursynth.imwri',
@@ -68,8 +66,9 @@ def is_venv() -> bool:
 def detect_vapoursynth_installation() -> str:
     try:
         spec = imputil.find_spec("vapoursynth")
+
         if spec is None:
-            raise ModuleNotFoundError()
+            raise ModuleNotFoundError
     except (ValueError, ModuleNotFoundError):
         print("Could not detect vapoursynth.")
         sys.exit(1)
@@ -81,9 +80,11 @@ def detect_vapoursynth_installation() -> str:
             print("The vapoursynth-module could not be found or imported.")
         else:
             return vapoursynth.__file__
+
     if spec.origin is None:
         print("VapourSynth's origin could not be determined.")
         sys.exit(1)
+
     return spec.origin
 
 
@@ -145,11 +146,16 @@ def get_vs_installation_site() -> str:
 is_64bits: bool = sys.maxsize > 2**32
 
 parser = argparse.ArgumentParser(description='A simple VapourSynth package manager')
-parser.add_argument('operation', choices=['install', 'update', 'upgrade', 'upgrade-all',
-                    'uninstall', 'installed', 'available', 'paths', "genstubs", "gendistinfo"])
+parser.add_argument(
+    'operation', choices=[
+        'install', 'update', 'upgrade', 'upgrade-all', 'uninstall',
+        'installed', 'available', 'paths', "genstubs", "gendistinfo"
+    ]
+)
 parser.add_argument('package', nargs='*', help='identifier, namespace or module to install, upgrade or uninstall')
-parser.add_argument('-f', action='store_true', dest='force',
-                    help='force upgrade for packages where the current version is unknown')
+parser.add_argument(
+    '-f', action='store_true', dest='force', help='force upgrade for packages where the current version is unknown'
+)
 parser.add_argument('-p', action='store_true', dest='portable', help='use paths suitable for portable installs')
 parser.add_argument('-d', action='store_true', dest='skip_deps', help='skip installing dependencies')
 parser.add_argument(
@@ -161,11 +167,12 @@ parser.add_argument('-b', dest='binary_path', help='custom binary install path')
 parser.add_argument('-s', dest='script_path', help='custom script install path')
 parser.add_argument("--stub-output-file", default="", help="Don't update the typestubs generated by vsrepo.")
 parser.add_argument("--force-dist-info", action="store_true", default=False, help="")
+
 args = parser.parse_args()
 
 is_64bits = args.target == 'win64'
 
-file_dirname: str = os.path.dirname(os.path.abspath(__file__))
+file_dirname = os.path.dirname(os.path.abspath(__file__))
 
 # VSRepo is installed to the site-packages.
 if os.path.abspath(file_dirname).startswith(os.path.abspath(sys.prefix)):
@@ -233,34 +240,6 @@ if not os.path.isfile(cmd7zip_path):
         cmd7zip_path = '7z.exe'
 
 installed_packages: Dict[str, str] = {}
-download_cache: Dict[str, bytearray] = {}
-
-
-def fetch_ur1(url: str, desc: Optional[str] = None) -> bytearray:
-    with urllib.request.urlopen(url) as urlreq:
-        if ('tqdm' in sys.modules) and (urlreq.headers['content-length'] is not None):
-            size = int(urlreq.headers['content-length'])
-            remaining = size
-            data = bytearray()
-            with tqdm.tqdm(total=size, unit='B', unit_scale=True, unit_divisor=1024, desc=desc) as t:
-                while remaining > 0:
-                    blocksize = min(remaining, 1024 * 128)
-                    data.extend(urlreq.read(blocksize))
-                    remaining = remaining - blocksize
-                    t.update(blocksize)
-            return data
-        else:
-            print('Fetching: ' + url)
-            return cast(bytearray, urlreq.read())
-
-
-def fetch_url_cached(url: str, desc: str = "") -> bytearray:
-    data = download_cache.get(url, None)
-    if data is None:
-        data = fetch_ur1(url, desc)
-        download_cache[url] = data
-    return data
-
 
 package_print_string = "{:25s} {:15s} {:11s} {:11s} {:s}"
 
@@ -799,25 +778,25 @@ def update_package_definition(url: str) -> None:
     except OSError:
         pass
 
-    localmtime = email.utils.formatdate(localmtimeval + 10, usegmt=True)
-    req_obj = urllib.request.Request(url, headers={'If-Modified-Since': localmtime, 'User-Agent': 'VSRepo'})
+    localmtime = formatdate(localmtimeval + 10, usegmt=True)
+    req_obj = Request(url, headers={'If-Modified-Since': localmtime, 'User-Agent': 'VSRepo'})
 
     try:
-        with urllib.request.urlopen(req_obj) as urlreq:
-            remote_modtime = email.utils.mktime_tz(email.utils.parsedate_tz(urlreq.info()['Last-Modified']))
+        with urlopen(req_obj) as urlreq:
+            remote_modtime = mktime_tz(parsedate_tz(urlreq.info()['Last-Modified']))
             data = urlreq.read()
             with zipfile.ZipFile(io.BytesIO(data), 'r') as zf:
                 with zf.open('vspackages3.json') as pkgfile:
                     with open(package_json_path, 'wb') as dstfile:
                         dstfile.write(pkgfile.read())
                     os.utime(package_json_path, times=(remote_modtime, remote_modtime))
-    except urllib.error.HTTPError as httperr:
+    except HTTPError as httperr:
         if httperr.code == 304:
-            print('Local definitions already up to date: ' + email.utils.formatdate(localmtimeval, usegmt=True))
+            print('Local definitions already up to date: ' + formatdate(localmtimeval, usegmt=True))
         else:
             raise
     else:
-        print('Local definitions updated to: ' + email.utils.formatdate(remote_modtime, usegmt=True))
+        print('Local definitions updated to: ' + formatdate(remote_modtime, usegmt=True))
 
 
 def get_vapoursynth_version() -> int:
@@ -948,94 +927,92 @@ def print_paths() -> None:
         print("Dist-Infos: <Will not be installed>")
 
 
-for name in args.package:
-    try:
-        assert isinstance(name, str)
-        get_package_from_name(name)
-    except Exception as e:
-        print(e)
-        sys.exit(1)
+def main() -> None:
 
-if args.operation == 'install':
-    detect_installed_packages()
-    rebuild_distinfo()
-
-    inst = (0, 0, 0)
     for name in args.package:
-        res = install_package(name)
-        inst = (inst[0] + res[0], inst[1] + res[1], inst[2] + res[2])
+        try:
+            assert isinstance(name, str)
+            get_package_from_name(name)
+        except Exception as e:
+            print(e)
+            sys.exit(1)
 
-    update_genstubs()
+    if args.operation == 'install':
+        detect_installed_packages()
+        rebuild_distinfo()
 
-    if (inst[0] == 0) and (inst[1] == 0):
-        print('Nothing done')
-    elif (inst[0] > 0) and (inst[1] == 0):
-        print('{} {} installed'.format(inst[0], 'package' if inst[0] == 1 else 'packages'))
-    elif (inst[0] == 0) and (inst[1] > 0):
-        print('{} missing {} installed'.format(inst[1], 'dependency' if inst[1] == 1 else 'dependencies'))
-    else:
-        print(
-            '{} {} and {} additional {} installed'.format(
-                inst[0],
-                'package' if inst[0] == 1 else 'packages', inst[1],
-                'dependency' if inst[1] == 1 else 'dependencies'))
-    if (inst[2] > 0):
-        print('{} {} failed'.format(inst[2], 'package' if inst[0] == 1 else 'packages'))
-elif args.operation in ('upgrade', 'upgrade-all'):
-    detect_installed_packages()
-    rebuild_distinfo()
-
-    inst = (0, 0, 0)
-    if args.operation == 'upgrade-all':
-        inst = upgrade_all_packages(args.force)
-    else:
+        inst = (0, 0, 0)
         for name in args.package:
-            res = upgrade_package(name, args.force)
+            res = install_package(name)
             inst = (inst[0] + res[0], inst[1] + res[1], inst[2] + res[2])
 
-    update_genstubs()
+        update_genstubs()
 
-    if (inst[0] == 0) and (inst[1] == 0):
-        print('Nothing done')
-    elif (inst[0] > 0) and (inst[1] == 0):
-        print('{} {} upgraded'.format(inst[0], 'package' if inst[0] == 1 else 'packages'))
-    elif (inst[0] == 0) and (inst[1] > 0):
-        print('{} missing {} installed'.format(inst[1], 'dependency' if inst[1] == 1 else 'dependencies'))
-    else:
-        print(
-            '{} {} upgraded and {} additional {} installed'.format(
-                inst[0],
-                'package' if inst[0] == 1 else 'packages', inst[1],
-                'dependency' if inst[1] == 1 else 'dependencies'))
-    if (inst[2] > 0):
-        print('{} {} failed'.format(inst[2], 'package' if inst[0] == 1 else 'packages'))
-elif args.operation == 'uninstall':
-    detect_installed_packages()
-    uninst = (0, 0)
-    for name in args.package:
-        uninst_res = uninstall_package(name)
-        uninst = (uninst[0] + uninst_res[0], uninst[1] + uninst_res[1])
-    if uninst[0] == 0:
-        print('No packages uninstalled')
-    else:
-        print('{} {} uninstalled'.format(uninst[0], 'package' if uninst[0] == 1 else 'packages'))
-    update_genstubs()
-elif args.operation == 'installed':
-    detect_installed_packages()
-    list_installed_packages()
-elif args.operation == 'available':
-    detect_installed_packages()
-    list_available_packages()
-elif args.operation == 'update':
-    update_package_definition('http://www.vapoursynth.com/vsrepo/vspackages3.zip')
-elif args.operation == 'paths':
-    print_paths()
-elif args.operation == "genstubs":
-    update_genstubs()
-elif args.operation == "gendistinfo":
-    detect_installed_packages()
-    rebuild_distinfo()
+        if (inst[0] == 0) and (inst[1] == 0):
+            print('Nothing done')
+        elif (inst[0] > 0) and (inst[1] == 0):
+            print('{} {} installed'.format(inst[0], 'package' if inst[0] == 1 else 'packages'))
+        elif (inst[0] == 0) and (inst[1] > 0):
+            print('{} missing {} installed'.format(inst[1], 'dependency' if inst[1] == 1 else 'dependencies'))
+        else:
+            print(
+                '{} {} and {} additional {} installed'.format(
+                    inst[0],
+                    'package' if inst[0] == 1 else 'packages', inst[1],
+                    'dependency' if inst[1] == 1 else 'dependencies'))
+        if (inst[2] > 0):
+            print('{} {} failed'.format(inst[2], 'package' if inst[0] == 1 else 'packages'))
+    elif args.operation in ('upgrade', 'upgrade-all'):
+        detect_installed_packages()
+        rebuild_distinfo()
 
+        inst = (0, 0, 0)
+        if args.operation == 'upgrade-all':
+            inst = upgrade_all_packages(args.force)
+        else:
+            for name in args.package:
+                res = upgrade_package(name, args.force)
+                inst = (inst[0] + res[0], inst[1] + res[1], inst[2] + res[2])
 
-def main() -> None:
-    ...
+        update_genstubs()
+
+        if (inst[0] == 0) and (inst[1] == 0):
+            print('Nothing done')
+        elif (inst[0] > 0) and (inst[1] == 0):
+            print('{} {} upgraded'.format(inst[0], 'package' if inst[0] == 1 else 'packages'))
+        elif (inst[0] == 0) and (inst[1] > 0):
+            print('{} missing {} installed'.format(inst[1], 'dependency' if inst[1] == 1 else 'dependencies'))
+        else:
+            print(
+                '{} {} upgraded and {} additional {} installed'.format(
+                    inst[0],
+                    'package' if inst[0] == 1 else 'packages', inst[1],
+                    'dependency' if inst[1] == 1 else 'dependencies'))
+        if (inst[2] > 0):
+            print('{} {} failed'.format(inst[2], 'package' if inst[0] == 1 else 'packages'))
+    elif args.operation == 'uninstall':
+        detect_installed_packages()
+        uninst = (0, 0)
+        for name in args.package:
+            uninst_res = uninstall_package(name)
+            uninst = (uninst[0] + uninst_res[0], uninst[1] + uninst_res[1])
+        if uninst[0] == 0:
+            print('No packages uninstalled')
+        else:
+            print('{} {} uninstalled'.format(uninst[0], 'package' if uninst[0] == 1 else 'packages'))
+        update_genstubs()
+    elif args.operation == 'installed':
+        detect_installed_packages()
+        list_installed_packages()
+    elif args.operation == 'available':
+        detect_installed_packages()
+        list_available_packages()
+    elif args.operation == 'update':
+        update_package_definition('http://www.vapoursynth.com/vsrepo/vspackages3.zip')
+    elif args.operation == 'paths':
+        print_paths()
+    elif args.operation == "genstubs":
+        update_genstubs()
+    elif args.operation == "gendistinfo":
+        detect_installed_packages()
+        rebuild_distinfo()
